Remove R_001 and R_002.
