+++ README.md
@@ -136,8 +136,6 @@ state the header and if the [pyhon behave](https://behave.readthedocs.io/en/late
 | R_011       | EST Endpoint for Onboarded Devices                              | [![Test](https://github.com/TrustPoint-Project/trustpoint/actions/workflows/r_011_feature_test.yml/badge.svg)](https://github.com/TrustPoint-Project/trustpoint/actions/workflows/r_011_feature_test.yml) |
 | R_012       | Language Selection and Translation                              | [![Test](https://github.com/TrustPoint-Project/trustpoint/actions/workflows/r_012_feature_test.yml/badge.svg)](https://github.com/TrustPoint-Project/trustpoint/actions/workflows/r_012_feature_test.yml) |
 | R_013       | Remote Credential Download                                      | [![Test](https://github.com/TrustPoint-Project/trustpoint/actions/workflows/r_013_feature_test.yml/badge.svg)](https://github.com/TrustPoint-Project/trustpoint/actions/workflows/r_013_feature_test.yml) |
-| R_101       | Devices are only allowed to communicate with valid certificates | [![Test](https://github.com/TrustPoint-Project/trustpoint/actions/workflows/r_101_feature_test.yml/badge.svg)](https://github.com/TrustPoint-Project/trustpoint/actions/workflows/r_101_feature_test.yml) |
-| R_102       | Encrypted Communication                                         | [![Test](https://github.com/TrustPoint-Project/trustpoint/actions/workflows/r_102_feature_test.yml/badge.svg)](https://github.com/TrustPoint-Project/trustpoint/actions/workflows/r_102_feature_test.yml) |
-| R_103       | Security Level Configuration                                    | [![Test](https://github.com/TrustPoint-Project/trustpoint/actions/workflows/r_103_feature_test.yml/badge.svg)](https://github.com/TrustPoint-Project/trustpoint/actions/workflows/r_103_feature_test.yml) |
-| R_104       | Certificate Template Security                                   | [![Test](https://github.com/TrustPoint-Project/trustpoint/actions/workflows/r_104_feature_test.yml/badge.svg)](https://github.com/TrustPoint-Project/trustpoint/actions/workflows/r_104_feature_test.yml) |
+| R_101       | Security Level Configuration                                    | [![Test](https://github.com/TrustPoint-Project/trustpoint/actions/workflows/r_101_feature_test.yml/badge.svg)](https://github.com/TrustPoint-Project/trustpoint/actions/workflows/r_101_feature_test.yml) |
+| R_102       | Certificate Template Security                                   | [![Test](https://github.com/TrustPoint-Project/trustpoint/actions/workflows/r_102_feature_test.yml/badge.svg)](https://github.com/TrustPoint-Project/trustpoint/actions/workflows/r_102_feature_test.yml) |
 | F_001       | NTEU must be able to execute R_001 and R_002.                   | [![Test](https://github.com/TrustPoint-Project/trustpoint/actions/workflows/f_001_feature_test.yml/badge.svg)](https://github.com/TrustPoint-Project/trustpoint/actions/workflows/f_001_feature_test.yml) |
