+++ trustpoint/features/steps/r_101_steps.py
@@ -1,77 +1,121 @@
 """Python steps file for R_101."""
 
-from enum import Enum
-
 from behave import given, runner, then, when
 
-Validity = Enum('Validity', ['valid', 'expired', 'revoked', 'self-signed', 'tampered'])
-Allow_Deny = Enum('Allow_Deny', ['allow', 'deny'])
 
-
-@given('the system enforces certificate validation for all devices')
-def step_given_system_enforces_cert_validation(context: runner.Context) -> None:  # noqa: ARG001
-    """Ensures that the system requires all devices to present valid certificates for communication.
+@given('the TrustPoint component {component} is selected')
+def step_given_component_selected(context: runner.Context, component: str) -> None:  # noqa: ARG001
+    """Simulates selecting a TrustPoint component in the security configuration panel.
 
     Args:
         context (runner.Context): Behave context.
+        component (str): The selected TrustPoint component.
     """
-    msg = 'STEP: Given the system enforces certificate validation for all devices'
+    msg = f'STEP: Given the TrustPoint component {component} is selected'
     raise AssertionError(msg)
 
 
-@given('a device has a {validity} certificate')
-def step_given_device_with_cert(context: runner.Context, validity: Validity) -> None:  # noqa: ARG001
-    """Simulates a device possessing a certificate.
+@given('the TrustPoint component {component} has security level {security_level}')
+def step_given_component_has_security_level(context: runner.Context, component: str, security_level: str) -> None:  # noqa: ARG001
+    """Ensures that the TrustPoint component has a specified security level.
 
     Args:
         context (runner.Context): Behave context.
-        validity (Validity): The validity.
+        component (str): The TrustPoint component.
+        security_level (str): The current security level.
     """
-    msg = f'STEP: Given a device has a {validity} certificate'
+    msg = f'STEP: Given the TrustPoint component {component} has security level {security_level}'
     raise AssertionError(msg)
 
 
-@given('a device does not present a certificate')
-def step_given_device_without_cert(context: runner.Context) -> None:  # noqa: ARG001
-    """Simulates a device attempting to communicate without providing a certificate.
+@when('the admin sets the security level to {security_level}')
+def step_when_admin_sets_security_level(context: runner.Context, security_level: str) -> None:  # noqa: ARG001
+    """Simulates an admin setting a security level for a component.
 
     Args:
         context (runner.Context): Behave context.
+        security_level (str): The new security level to be set.
     """
-    msg = 'STEP: Given a device does not present a certificate'
+    msg = f'STEP: When the admin sets the security level to {security_level}'
     raise AssertionError(msg)
 
 
-@when('the device attempts to establish communication')
-def step_when_device_attempts_communication(context: runner.Context) -> None:  # noqa: ARG001
-    """Simulates a device attempting to establish communication with the system.
+@when('the system is restarted')
+def step_when_system_restarts(context: runner.Context) -> None:  # noqa: ARG001
+    """Simulates restarting the system.
 
     Args:
         context (runner.Context): Behave context.
     """
-    msg = 'STEP: When the device attempts to establish communication'
+    msg = 'STEP: When the system is restarted'
     raise AssertionError(msg)
 
 
-@then('the system should {allow_deny} the communication')
-def step_then_allow_deny_communication(context: runner.Context, allow_deny: Allow_Deny) -> None:  # noqa: ARG001
-    """Ensures that the system allows/denies communication when a valid certificate is presented.
+@when('an unauthorized user attempts access')
+def step_when_unauthorized_access_attempted(context: runner.Context) -> None:  # noqa: ARG001
+    """Simulates an unauthorized user attempting to access a system component.
 
     Args:
         context (runner.Context): Behave context.
-        allow_deny (Allow_Deny): allow or deny.
     """
-    msg = f'STEP: Then the system should {allow_deny} the communication'
+    msg = 'STEP: When an unauthorized user attempts access'
     raise AssertionError(msg)
 
 
-@then('log the authentication failure with reason {reason}')
-def step_then_log_failure(context: runner.Context, reason: str) -> None:  # noqa: ARG001
-    """Ensures that the system logs authentication failures with the appropriate reason.
+@then('the system should apply the security level {security_level}')
+def step_then_system_applies_security_level(context: runner.Context, security_level: str) -> None:  # noqa: ARG001
+    """Ensures that the system applies the specified security level.
 
     Args:
         context (runner.Context): Behave context.
-        reason (str): The reason for authentication failure.
+        security_level (str): The expected security level applied.
     """
-    msg = f'STEP: Then log the authentication failure with reason {reason}'
+    msg = f'STEP: Then the system should apply the security level {security_level}'
+    raise AssertionError(msg)
+
+
+@then('the system should reject the input with error {error_message}')
+def step_then_reject_invalid_input(context: runner.Context, error_message: str) -> None:  # noqa: ARG001
+    """Ensures that an invalid security level input is rejected with an error.
+
+    Args:
+        context (runner.Context): Behave context.
+        error_message (str): The expected error message.
+    """
+    msg = f'STEP: Then the system should reject the input with error {error_message}'
+    raise AssertionError(msg)
+
+
+@then('the TrustPoint component {component} should still have security level {security_level}')
+def step_then_security_level_persists(context: runner.Context, component: str, security_level: str) -> None:  # noqa: ARG001
+    """Ensures that the security level persists after a system restart.
+
+    Args:
+        context (runner.Context): Behave context.
+        component (str): The TrustPoint component.
+        security_level (str): The expected security level after restart.
+    """
+    msg = f'STEP: Then the TrustPoint component {component} should still have security level {security_level}'
+    raise AssertionError(msg)
+
+
+@then('access should be denied')
+def step_then_access_denied(context: runner.Context) -> None:  # noqa: ARG001
+    """Ensures that access is denied based on security level settings.
+
+    Args:
+        context (runner.Context): Behave context.
+    """
+    msg = 'STEP: Then access should be denied'
+    raise AssertionError(msg)
+
+
+@then('the system should log the security level change with details')
+def step_then_log_security_change(context: runner.Context) -> None:  # noqa: ARG001
+    """Ensures that changes to security levels are logged.
+
+    Args:
+        context (runner.Context): Behave context.
+    """
+    msg = 'STEP: Then the system should log the security level change with details'
     raise AssertionError(msg)
